app/src/main/java/com/zekierciyas/takeanote/data/db/NotesDao.kt
package com.zekierciyas.takeanote.data.db

import androidx.lifecycle.LiveData
import androidx.room.*
import com.zekierciyas.takeanote.models.Note

@Dao
interface NotesDao {

    @Query("SELECT * FROM notes_table ORDER by date DESC")
    fun getAllNotes(): LiveData<List<Note>>

    @Insert(onConflict = OnConflictStrategy.IGNORE)
    suspend fun insertNote(note: Note)

    @Update
    suspend fun updateNote(note: Note)

    @Delete
    suspend fun deleteNote(note: Note)

    @Query("DELETE FROM notes_table")
    suspend fun deleteAllNotes()

}